moved FramedDataset tests to their own module
